View3D is now exported to be available from plugins
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/core/trunk@312 77d55b08-817d-0410-979b-1fec1beb8c2e

+++ cluster/SimpleClusterView.cpp
@@ -67,7 +67,7 @@ namespace intern {
         vislib::net::TIPHostEntry *he = static_cast<vislib::net::TIPHostEntry*>(userContext);
         vislib::Array<vislib::net::IPAgnosticAddress> a1 = he->GetAddresses();
         vislib::net::NetworkInformation::UnicastAddressList a2 = adapter.GetUnicastAddresses();
-        for (int i2 = 0; i2 < a2.Count(); i2++) {
+        for (SIZE_T i2 = 0; i2 < a2.Count(); i2++) {
             if (a1.Contains(a2[i2].GetAddress())) {
                 return true;
             }
